Renamed parentheses token classes, added unit tests for new class tokens.
