+++ src/Kingsland.MofParser.UnitTests/Lexer/Tokens/attributeclose.mof
@@ -1,0 +1,1 @@
+]
