+++ src/Kingsland.MofParser.UnitTests/Lexer/Tokens/attributeopen.mof
@@ -1,0 +1,1 @@
+[
