+++ src/Kingsland.MofParser.UnitTests/Lexer/Tokens/colon.mof
@@ -1,0 +1,1 @@
+:
