+++ src/Kingsland.MofParser.UnitTests/Lexer/Tokens/equalsoperator.json
