+++ src/Kingsland.MofParser.UnitTests/Lexer/Tokens/parenthesesclose.json
@@ -1,0 +1,18 @@
+{
+  "$type": "System.Collections.Generic.List`1[[Kingsland.MofParser.Tokens.Token, Kingsland.MofParser]], mscorlib",
+  "$values": [
+    {
+      "$type": "Kingsland.MofParser.Tokens.ParenthesesCloseToken, Kingsland.MofParser",
+      "Extent": {
+        "$type": "Kingsland.MofParser.Lexing.SourceExtent, Kingsland.MofParser",
+        "StartPosition": 0,
+        "StartLineNumber": 1,
+        "StartColumnNumber": 1,
+        "EndPosition": 0,
+        "EndLineNumber": 1,
+        "EndColumnNumber": 1,
+        "Text": ")"
+      }
+    }
+  ]
+}
