+++ src/Kingsland.MofParser.UnitTests/Lexer/Tokens/parenthesesclose.mof
@@ -1,0 +1,1 @@
+)
