+++ src/Kingsland.MofParser.UnitTests/Lexer/Tokens/parenthesesopen.mof
@@ -1,0 +1,1 @@
+(
