+++ src/Kingsland.MofParser.UnitTests/Lexer/Tokens/pragma_lower.mof
@@ -1,0 +1,1 @@
+#pragma
