+++ src/Kingsland.MofParser.UnitTests/Lexer/Tokens/statementend.mof
@@ -1,0 +1,1 @@
+;
