+++ src/Kingsland.MofParser/Kingsland.MofParser.csproj
@@ -60,7 +60,7 @@
     <Compile Include="Lexing\LexerStreamBase.cs" />
     <Compile Include="Tokens\AttributeCloseToken.cs" />
     <Compile Include="Tokens\AttributeOpenToken.cs" />
-    <Compile Include="Tokens\CloseParenthesesToken.cs" />
+    <Compile Include="Tokens\ParenthesesCloseToken.cs" />
     <Compile Include="Tokens\ColonToken.cs" />
     <Compile Include="Tokens\CommaToken.cs" />
     <Compile Include="Tokens\BooleanLiteralToken.cs" />
@@ -76,7 +76,7 @@
     <Compile Include="Tokens\BlockCloseToken.cs" />
     <Compile Include="Tokens\IntegerLiteralToken.cs" />
     <Compile Include="Tokens\NullLiteralToken.cs" />
-    <Compile Include="Tokens\OpenParenthesesToken.cs" />
+    <Compile Include="Tokens\ParenthesesOpenToken.cs" />
     <Compile Include="Tokens\PragmaToken.cs" />
     <Compile Include="Tokens\StatementEndToken.cs" />
     <Compile Include="Tokens\StringLiteralToken.cs" />
